+++ src/main/webapp/entries.js
@@ -53,10 +53,10 @@ module.exports = {
     "./resources/js/pages/projects/metadata-template/create-metadata-template.js",
   "visualizations-phylogenetics":
     "./resources/js/pages/visualizations/phylogenetics/index.js",
-  "projects-associated-edit":
-    "./resources/js/pages/projects/associated-projects/edit.module.js",
   "project-users-groups":
     "./resources/js/pages/projects/project-users-groups.js",
+  "associated-projects":
+    "./resources/js/pages/projects/associated-projects/index.js",
   "sample-files": "./resources/js/pages/samples/sample-files.js",
   "project-reference-files":
     "./resources/js/pages/projects/project-reference-files.js",