+++ src/main/webapp/resources/js/pages/cart/components/CartSamples.jsx
@@ -14,7 +14,7 @@ import {
 import AutoSizer from "react-virtualized-auto-sizer";
 import { Button, Input } from "antd";
 import { FixedSizeList as VList } from "react-window";
-import { getI18N } from "../../../utilities/i18n-utilties";
+import { getI18N } from "../../../utilities/i18n-utilities";
 import { actions } from "../../../redux/reducers/cart";
 import { sampleDetailsActions } from "../../../components/SampleDetails";
 import { SampleRenderer } from "./SampleRenderer";